requirements.txt
python-binance
ta
pandas
schedule
requests
